webapp/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
jinja2
python-multipart
python-dotenv
pydantic[email]
cryptography